requirements.txt
# Server
fastapi
fastapi_pagination>=0.12.28
uvicorn
aiofiles
asgiref
python-multipart
tortoise-orm[asyncpg]==0.23.0
aerich==0.8.2
vecs==0.4.3
# Libraries for Python
httpx>=0.28.1,<1.0.0
python-dotenv
asynctempfile
Pillow==10.4.0
types-Pillow
pandas
openpyxl
pdf2image
cairosvg
aiocache
# Authentication
google-auth>=2.28.0
pyjwt>=2.8.0
# SaaS
sentry-sdk[fastapi]
aioboto3
types-aioboto3[essential]
openai==1.35.7
deepgram_sdk==3.3.1
groq==0.9.0
# tools
pre-commit
pyright
autopep8
mypy
python-pptx>=0.6.21
python-dotenv>=1.0.0
asyncpg==0.30.0
moviepy
numpy
requests
scenedetect[opencv]
PyMuPDF

google-genai
pydantic==2.10.6
tomlkit
qdrant-client
vertexai

langgraph